Document the bootstrap procedure
As the module graph analyzer is written in TypeScript, its module graph
must be pre-generated or it would be impossible to run within Nix. The
bootstrapping procedure involves building a vendor directory with `deno
vendor`, importing that into the Nix store, and running that on the
original source. For more information, see `src/boostrap.md`.