Handle JSON imports
Due to the fact Deno doesn't store MIME information in the lock file,
there is a manual exception where .json files imported from a url with
the "?module" query parameter (used by esm.sh) are assumed to actually
be wrapped JS objects. For these, the file extension is changed to .js
while building the vendor directory.